(refactoring) UT helpers lib KIKIMR-20902
